requirements.txt
blinker==1.4
click==6.6
Flask==0.11.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-openid==2.2.5
SQLAlchemy==1.0.15
Werkzeug==0.11.11
Whoosh==2.7.4
WTForms==2.1
sqlalchemy-migrate==0.9.2